Rename: active brown box를 공통 컴포넌트로 변경
