+++ src/components/common/ActiveBrownBox/index.tsx
