+++ src/components/features/Layout/ProfileQuestionLayout/QuestionList/index.tsx
@@ -2,7 +2,7 @@ import { useEffect } from 'react'
 
 import { Flex } from '@chakra-ui/react'
 
-import { ActiveBrownBox } from '@/components/features/Layout/ActiveBrownBox'
+import { ActiveBrownBox } from '@/components/common/ActiveBrownBox'
 import { useSeletedQuestionStore } from '@/stores/selected-question'
 import { QuestionItem } from '@/types'
